Tax rate authorization

+++ Modules/Tax/Http/Controllers/TaxRateController.php
@@ -13,6 +13,7 @@ use Modules\Core\Http\Controllers\BaseController;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Illuminate\Validation\ValidationException;
 use Modules\Country\Entities\Region;
+use Modules\Tax\Exceptions\TaxRateCanNotBeDeleted;
 
 class TaxRateController extends BaseController
 {
@@ -23,8 +24,11 @@ class TaxRateController extends BaseController
         $this->repository = $taxRateRepository;
         $this->model = $taxRate;
         $this->model_name = "Tax Rate";
+        $exception_statuses = [
+            TaxRateCanNotBeDeleted::class => 403
+        ];
 
-        parent::__construct($this->model, $this->model_name);
+        parent::__construct($this->model, $this->model_name, $exception_statuses);
     }
 
     public function collection(object $data): ResourceCollection
@@ -108,6 +112,8 @@ class TaxRateController extends BaseController
     {
         try
         {
+            $fetched = $this->model->find($id);
+            if($fetched->tax_rules->count() != 0) throw new TaxRateCanNotBeDeleted(__("core::app.response.delete-failed", ["name" => $this->model_name]));
             $this->repository->delete($id);
         }
         catch( Exception $exception )
